requirements.txt
tornado==5.1
mysqlclient==1.3.13
requests==2.18.4
bs4==0.0.1
nltk==3.2.4
pynacl==1.2.1
